use '273 registers for ucode, initial PC design
The microcode registers need to be reset to a known
state on startup.  The 74HCT273 is appropriate for this.
One consequence is that all microcode signals must be
designed such that a 0 is the correct output for
coming out of reset: the main thing that needs to
happen is putting an initial value into the program
counter.  I'm assuming that a '541 buffer will
put an appropriate byte value (maybe 0xC0) on the
D bus on reset, and the PC can load this in both
bytes (so, effective initial PC of 0xC0C0).

Created an initial design for the program counter.
Used 74HC191 synchronous counters, with a 74HC10
triple 3-input NAND to enable counting for each 4
bits as appopriate based on the terminal count status
of the prior counters.